+++ src/pages/Home/HomePage.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useRecoilValue } from 'recoil';
-import HomePageExternal from './HomePageExternal';
-import HomePageInternal from './HomePageInternal';
+import HomePageExternal from './HomePageExternal/HomePageExternal';
+import HomePageInternal from './HomePageInternal/HomePageInternal';
 
 import { userInfoAtom } from '../../services/recoil/userInfo';
