+++ src/pages/Home/HomePageExternal/CommunityCarousel.tsx
